backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai>=0.8.0
python-dotenv>=1.0.1
langchain>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.20
pypdf>=4.0.0
chromadb>=0.4.0
langchain-community>=0.0.10

